+++ CHANGES.md
@@ -1,2 +1,7 @@
+5.0.0
+ * Migrated to new CSPROJ project system
+ * Updated dependencies: WindowsAzure.Storage 8.6.0, Serilog 2.6.0, Serilog.Sinks.PeriodicBatching 2.1.1
+ * Fix #36 - Allow using SAS URI for logging.
+
 1.5
  * Moved from serilog/serilog